Allow more complicated post-pip-install setup sequences (#8)
There's always been an option to run arbitrary code after the pip install step in order to set up the Python venv. However, that code could not use any other files from the current repository (so it was only useful for, e.g., downloading and installing something). This PR adds a new option, run_after_pip_srcs, which allows the user to specify a set of files which will be available to their post-install script.

I am using this to build + install a C++ module inside the venv.