Escape arguments to the coverage script (#9)
We pass to the coverage report-generating script strings which specify where to find certain source files. We want to support providing multiple such files, hence we want each argument to include spaces. I thought it would be enough to just pass them as two separate arguments in the sh_binary rule, but it seems it's necessary to actually escape the spaces. This PR addresses that.